rename eia_bulk_data to eia_bulk_elec; move two funcs to transform
